% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_data.R
\name{updating_data}
\alias{updating_data}
\title{updated information of corona}
\usage{
updating_data()
}
\description{
Provides tidy data of corona outbreak.output is data frame and no input required.
}
\examples{
updated_data = updating_data() #returns a tidy data frame of corona outbreak data
}
